Add RoomList for search result and my room page
